fix(db): update reblooms table to support 64-bit integer IDs
Changed 'id' to BIGSERIAL and 'bloom_id' to BIGINT to prevent
NumericValueOutOfRange errors. This ensures the reblooms table
can correctly reference the large timestamp-based IDs used in
the blooms table.

+++ db/schema.sql
@@ -28,8 +28,8 @@ CREATE TABLE hashtags (
 );
 
 CREATE TABLE reblooms (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     rebloomer_id INTEGER NOT NULL REFERENCES users(id),
-    bloom_id INTEGER NOT NULL REFERENCES blooms(id),
+    bloom_id BIGINT NOT NULL REFERENCES blooms(id),
     rebloom_timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
